pyproject.toml
[tool.poetry]
name = "kochbuch"
version = "0.1.0"
description = ""
authors = ["Daniel Wiegand <daniel.a.wiegand@posteo.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
shiny = "^0.5.1"
shinywidgets = "^0.2.1"
ipykernel = "^6.25.2"
ipywidgets = "^8.1.1"
numpy = "^1.26.0"
pandas = "^2.1.1"
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
ipyregulartable = "^0.2.1"
qgrid = "^1.3.1"
ipyleaflet = "^0.17.4"
ipydatagrid = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
